feat(EIDOMNI-680): refactor RenewalRequestDto and RenewalResponseDto package structure and update imports
