+++ issuer-application/src/main/java/ch/admin/bj/swiyu/issuer/infrastructure/config/OpenApiConfig.java
@@ -3,8 +3,8 @@ package ch.admin.bj.swiyu.issuer.infrastructure.config;
 import ch.admin.bj.swiyu.issuer.domain.openid.metadata.*;
 import ch.admin.bj.swiyu.issuer.dto.callback.WebhookCallbackDto;
 import ch.admin.bj.swiyu.issuer.dto.exception.ApiErrorDto;
-import ch.admin.bj.swiyu.issuer.service.renewal.RenewalRequestDto;
-import ch.admin.bj.swiyu.issuer.service.renewal.RenewalResponseDto;
+import ch.admin.bj.swiyu.issuer.dto.renewal.RenewalRequestDto;
+import ch.admin.bj.swiyu.issuer.dto.renewal.RenewalResponseDto;
 import io.swagger.v3.core.converter.ModelConverters;
 import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
 import io.swagger.v3.oas.annotations.security.SecurityScheme;
