+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/dto/renewal/RenewalRequestDto.java
@@ -1,4 +1,4 @@
-package ch.admin.bj.swiyu.issuer.service.renewal;
+package ch.admin.bj.swiyu.issuer.dto.renewal;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
